pyproject.toml
[tool.poetry]
name = "django_q_sqectacular"
version = "0.1.0"
description = ""
authors = ["Benedict Thekkel <bthekkel1@gmail.com>"]
readme = "README.md"
package-mode = false


[tool.poetry.dependencies]
python = "^3.12"
asgiref = "3.7.2"
certifi = "2022.12.7"
cffi = "1.16.0"
charset-normalizer = "3.0.1"
crispy-bootstrap5 = "2023.10"
cryptography = "39.0.1"
defusedxml = "0.7.1"
django = "5.0.3"
django-allauth = "0.61.1"
django-crispy-forms = "2.1"
django-debug-toolbar = "4.2.0"
idna = "3.4"
oauthlib = "3.2.2"
packaging = "23.1"
psycopg = "3.1.18"
psycopg-binary = "3.1.18"
pycparser = "2.21"
pyjwt = "2.6.0"
python3-openid = "3.2.0"
requests = "2.28.2"
requests-oauthlib = "1.3.1"
sqlparse = "0.4.3"
typing-extensions = "4.9.0"
urllib3 = "1.26.14"
whitenoise = "6.6.0"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.9"
django-environ = "^0.11.2"
pytest = "^8.3.1"
pytest-django = "^4.8.0"
faker = "^26.0.0"
factory-boy = "^3.3.0"
django-polymorphic = "^3.1.0"
setuptools = "^72.2.0"
boto3 = "^1.34.162"
python-dotenv = "^1.0.1"
croniter = "^3.0.3"
django-object-actions = "^4.2.0"
django-post-office = "^3.9.0"
jinja2 = "^3.1.4"
django-extensions = "^3.2.3"
dj-database-url = "^2.2.0"
gunicorn = "^23.0.0"
django-phonenumber-field = {extras = ["phonenumberslite"], version = "^8.0.0"}
django-annoying = "^0.10.7"
freezegun = "^1.5.1"
django-q2 = "^1.7.2"
redis = "^5.0.8"
mkdocs = "^1.6.1"
mkdocs-material = "^9.5.39"
mkdocstrings = {extras = ["python"], version = "^0.26.1"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
